refactor(BSVHU/BSDA): factorize canAccessDraftOrgIds
